requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-dotenv==1.0.0
sqlalchemy[asyncio]~=2.0.21
redis~=5.0.0
pydantic~=2.3.0
alembic~=1.12.0
passlib[bcrypt]~=1.7.4
PyJWT~=2.8.0
starlette~=0.27.0
APScheduler~=3.10.4
asyncio~=3.4.3